requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.2.0
qrcode==7.4.2
Pillow==10.4.0
bcrypt==4.1.2
python-dotenv==1.0.0
PyJWT==2.8.0

